+++ src/WebApiBolierplate/API/Controllers/CustomBaseController.cs
@@ -1,5 +1,4 @@
-﻿using API.Helpers;
-using Core.Constants;
+﻿using Core.Constants;
 using Microsoft.AspNetCore.Mvc;
 
 namespace API.Controllers;
@@ -8,12 +7,10 @@ namespace API.Controllers;
 [ApiController]
 public class CustomBaseController : ControllerBase
 {
-    protected readonly WebAPIHelper webAPIHelper;
     protected string loginId => User.FindFirst(CustomClaims.UserIdentifier)?.Value;
     protected string companyId => User.FindFirst(CustomClaims.CompanyIdentifier).Value;
 
     public CustomBaseController()
     {
-        webAPIHelper = new WebAPIHelper();
     }
 }