+++ CHANGELOG.md
@@ -4,9 +4,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 ## [Unreleased]
 ### Added
- - Added interfaces for library classes
+ - Added interfaces for library classes, helper functions
 
 ### Changed
+ - Helper functions now use dependency injection
  - Made all non-derived classes sealed for security and performance
 
 ## [7.0.0] - 2023-02-27